Change Count value to long

+++ src/main/java/com/ub/core/picture/services/PictureColorService.java
@@ -45,7 +45,7 @@ public class PictureColorService {
                 searchItemColorAdminRequest.getCurrentPage(),
                 searchItemColorAdminRequest.getPageSize(),
                 result);
-        searchItemColorAdminResponse.setAll(count.intValue());
+        searchItemColorAdminResponse.setAll(count);
         searchItemColorAdminResponse.setQuery(searchItemColorAdminRequest.getQuery());
         return searchItemColorAdminResponse;
     }
